requirements.txt
fastapi
uvicorn
pydantic
python-multipart
google-cloud-speech
google-genai
google-cloud-firestore
google-cloud-storage
pytest
ruff
websockets==13.0
pydub
audioop-lts
dotenv
elevenlabs
openai>=1.0.0